Added common config-file in booster
* The config file are to collect all common configurations for the booster.
    - Currently the common functions are moved; in the next couple of commits more common functionality will be moved.

* The overall goal is consolidate the functionality and reduce the amount of repeated code.

NOTE: All tests passes locally.
