+++ src/booster/mod.rs
@@ -2,3 +2,4 @@ pub mod booster;
 pub mod multi_output;
 pub mod predict;
 pub mod setters;
+pub mod config;
