+++ src/booster/predict.rs
@@ -1,4 +1,4 @@
-use super::booster::ContributionsMethod;
+use crate::booster::config::ContributionsMethod;
 use crate::{
     objective_functions::Objective, shapley::predict_contributions_row_shapley, tree::tree::Tree, utils::odds, Matrix,
     PerpetualBooster,
