+++ src/booster/setters.rs
@@ -1,4 +1,4 @@
-use super::booster::MissingNodeTreatment;
+use crate::booster::config::MissingNodeTreatment;
 use crate::{constraints::ConstraintMap, objective_functions::Objective, PerpetualBooster};
 use std::collections::HashSet;
